research/clickhouse/requirements.txt
clickhouse-driver==0.2.6
fake==0.8
docker==6.1.3